Rename some GF array modules
