+++ galois/array_2.py
