+++ galois/array_2m.py
