+++ galois/array_p.py
