+++ tests/fields/test_properties.py
@@ -7,15 +7,6 @@ import pytest
 import galois
 
 
-def test_properties(properties):
-    GF = properties["GF"]
-    assert GF.characteristic == properties["characteristic"]
-    assert GF.degree == properties["degree"]
-    assert GF.order == properties["order"]
-    assert GF.alpha == properties["alpha"]
-    assert all(GF.prim_poly.coeffs == properties["prim_poly"])
-
-
 def test_characteristic(field):
     if field.order < 2**16:
         a = field.Elements()
